pyproject.toml
[project]
name = "multiagent-prompt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-generativeai>=0.8.5",
    "openai-agents>=0.1.0",
    "py-dotenv>=0.1",
    "requests>=2.32.4",
]
